[artf7711] Expose maximumRetryCount and maximumWaitingTime (core settings) properties in ccfcoredefaultconfig.xml
LandscapeCCFPropertiesController- Whenever the core gets upgraded with newly direction config(dc) properties, Web UI was not able to display connector properties and it failed with no entity found for query exception.

If the newly added dc name is not found by findDirectionConfigsByDirectionAndName() it throws the above error. Hence added the condition to check the result list is empty on executing the query; if empty add the config  to the default properties list.

Added the same logic while saving the newly added DC.


git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/branches/CCF%202.2@5555 6771e1e2-0f46-0410-b5d7-d846ff89f743