[artf7694] Disable Test Connection for third party integration
Added new property displayTestConnection in GenericParticipantConfigItemFactory

Added the condition to the participant settings controller to 
display test connection  based on the displayTestConnection  
property value.

If the customer for third party integration is willing to do an implementation for testing connection he could enable this property and set the value to true. Also implement the corresponding test connection implementation in the class. This would enable test connection button in UI.

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/branches/CCF%202.2@5565 6771e1e2-0f46-0410-b5d7-d846ff89f743