[artf7961]: core version and description are missing in the connector upgrade [artf7960]: Connector behaviors failed to load the properties of JIRA to TeamForge(Reverse)
Changes done:

ccfcoredefaultconfig.xml - added in all the plugins build xml e.g. build-JIRA-ccfcore.xml
CreateLandscapeHelper.java - Creation of landscape is failed, so did a vaildation whether max attachment size is already created or not.
TFSGenericParticipantConfigValidator.java - Validation message is not correctly displayed and it throws exception, so fixed it.

git-svn-id: https://ctf.open.collab.net/svn/repos/ccf/branches/CCF%202.2@5678 6771e1e2-0f46-0410-b5d7-d846ff89f743