Encode password to support special characters in URI
Encoded password with encodeURIComponent() in doAjaxPort function for
both QC and TF direction. This will allow to include special characters
(like '%') in password and to support the same in data URI so that the
test connection won't throw any (error:object) exception even if it
contains special characters.

Added encodeURIComponent() for all the ajax post parameters in the
tfsettings,participantsettings and createlandscape.